Mostly fixed, todo ilist conversion
Seems to be mostly fixed, an issue with the Unsorted dropzone using IList<AnswerItem> instead of List<AnswerItem>. IList is used for the shuffler, but I'm not sure if I can shuffle them without IList. It's definitely worth investigating getting rid of IList because just dealing with one type would be nice.